+++ npm/__tests__/features/plugins/limit/limitReachedEvent.js
@@ -1,11 +1,11 @@
-/* global given, delegate, abnegate */
+/* global given */
 
 import Cocooned from '@notus.sh/cocooned/cocooned'
 import { jest } from '@jest/globals'
 import { setup, clickEvent } from '@cocooned/tests/support/helpers'
 import { getItems, getAddLink } from '@cocooned/tests/support/selectors'
 
-import itBehavesLikeAnEventListener from '@cocooned/tests/unit/shared/events/listener'
+import itBehavesLikeAnEventListener from '@cocooned/tests/shared/events/listener'
 
 describe('A Cocooned setup', () => {
   given('template', () => `
@@ -29,21 +29,18 @@ describe('A Cocooned setup', () => {
   beforeEach(() => setup(document, given))
 
   describe('when limit is reached', () => {
-    beforeEach(() => delegate('cocooned:limit-reached', ['event', 'node', 'cocooned']))
-    afterEach(() => abnegate('cocooned:limit-reached'))
-
     describe('a coconned:limit-reached event', () => {
       it('is triggered', () => {
         const listener = jest.fn()
-        given.container.addEventListener('$cocooned:limit-reached', listener)
+        given.container.addEventListener('cocooned:limit-reached', listener)
         given.addLink.dispatchEvent(clickEvent())
 
         expect(listener).toHaveBeenCalled()
-        given.container.removeEventListener('$cocooned:limit-reached', listener)
+        given.container.removeEventListener('cocooned:limit-reached', listener)
       })
 
       itBehavesLikeAnEventListener({
-        listen: (listener) => given.container.addEventListener('$cocooned:limit-reached', listener),
+        listen: (listener) => given.container.addEventListener('cocooned:limit-reached', listener),
         dispatch: () => given.addLink.dispatchEvent(clickEvent())
       })
     })