+++ npm/__tests__/features/removeLink/removeEvents.js
@@ -1,12 +1,12 @@
-/* global given, delegate, abnegate */
+/* global given */
 
 import Cocooned from '@notus.sh/cocooned/cocooned'
 import { jest } from '@jest/globals'
 import { setup, clickEvent } from '@cocooned/tests/support/helpers'
 import { getItem, getRemoveLink } from '@cocooned/tests/support/selectors'
 
-import itBehavesLikeAnEventListener from '@cocooned/tests/unit/shared/events/listener'
-import itBehavesLikeACancellableEvent from '@cocooned/tests/unit/shared/events/cancelable'
+import itBehavesLikeAnEventListener from '@cocooned/tests/shared/events/listener'
+import itBehavesLikeACancellableEvent from '@cocooned/tests/shared/events/cancelable'
 
 describe('A Cocooned setup', () => {
   given('template', () => `
@@ -34,26 +34,17 @@ describe('A Cocooned setup', () => {
   beforeEach(() => setup(document, given))
 
   describe('events on remove', () => {
-    beforeEach(() => {
-      delegate('cocooned:before-remove', ['event', 'node', 'cocooned'])
-      delegate('cocooned:after-remove', ['event', 'node', 'cocooned'])
-    })
-    afterEach(() => {
-      abnegate('cocooned:before-remove')
-      abnegate('cocooned:after-remove')
-    })
-
     describe('a before-remove event', () => {
       it('is triggered', () => {
         const listener = jest.fn()
-        given.container.addEventListener('$cocooned:before-remove', listener)
+        given.container.addEventListener('cocooned:before-remove', listener)
         given.removeLink.dispatchEvent(clickEvent())
 
         expect(listener).toHaveBeenCalled()
       })
 
       itBehavesLikeAnEventListener({
-        listen: (listener) => given.container.addEventListener('$cocooned:before-remove', listener),
+        listen: (listener) => given.container.addEventListener('cocooned:before-remove', listener),
         dispatch: () => given.removeLink.dispatchEvent(clickEvent())
       })
     })
@@ -61,22 +52,21 @@ describe('A Cocooned setup', () => {
     describe('an after-remove event', () => {
       it('is triggered', () => {
         const listener = jest.fn()
-        given.container.addEventListener('$cocooned:after-remove', listener)
+        given.container.addEventListener('cocooned:after-remove', listener)
         given.removeLink.dispatchEvent(clickEvent())
 
         expect(listener).toHaveBeenCalled()
       })
 
       itBehavesLikeAnEventListener({
-        listen: (listener) => given.container.addEventListener('$cocooned:after-remove', listener),
+        listen: (listener) => given.container.addEventListener('cocooned:after-remove', listener),
         dispatch: () => given.removeLink.dispatchEvent(clickEvent())
       })
     })
 
     itBehavesLikeACancellableEvent({
       event: 'remove',
-      dispatch: () => { given.removeLink.dispatchEvent(clickEvent()) },
-      trigger: () => { $(given.removeLink).trigger('click') }
+      dispatch: () => { given.removeLink.dispatchEvent(clickEvent()) }
     })
   })
 })