+++ npm/__tests__/integration/cocoon/events.js
@@ -1,4 +1,4 @@
-/* global given, delegate, abnegate */
+/* global given */
 
 import Cocooned from '@notus.sh/cocooned/cocooned'
 import { jest } from '@jest/globals'
@@ -31,12 +31,9 @@ describe('A Cocoon setup using Cocoon events', () => {
   beforeEach(() => setup(document, given))
 
   describe('when add link is clicked', () => {
-    beforeEach(() => delegate('cocoon:before-insert'))
-    afterEach(() => abnegate('cocoon:before-insert'))
-
     it('fires a before-insert event', () => {
       const listener = jest.fn()
-      given.container.addEventListener('$cocoon:before-insert', listener)
+      given.container.addEventListener('cocoon:before-insert', listener)
       given.addLink.dispatchEvent(clickEvent())
 
       expect(listener).toHaveBeenCalled()
@@ -47,14 +44,11 @@ describe('A Cocoon setup using Cocoon events', () => {
     beforeEach(() => given.addLink.dispatchEvent(clickEvent()))
 
     describe('when remove link is clicked', () => {
-      beforeEach(() => delegate('cocoon:before-remove'))
-      afterEach(() => abnegate('cocoon:before-remove'))
-
       given('removeLink', () => getRemoveLink(given.container))
 
       it('fires a before-remove event', () => {
         const listener = jest.fn()
-        given.container.addEventListener('$cocoon:before-remove', listener)
+        given.container.addEventListener('cocoon:before-remove', listener)
         given.removeLink.dispatchEvent(clickEvent())
 
         expect(listener).toHaveBeenCalled()