+++ npm/__tests__/shared/events/cancelable.js
@@ -1,0 +1,31 @@
+/* global given, $ */
+/* eslint jest/no-export: "off" -- This is a shared examples */
+
+import { jest } from '@jest/globals'
+
+export default ({ event, dispatch }) => {
+  describe('a cancelable operation', () => {
+    it('is canceled if default behavior of the before event is prevented', () => {
+      const canceler = jest.fn(e => e.preventDefault())
+      given.container.addEventListener(`cocooned:before-${event}`, canceler)
+
+      const listener = jest.fn()
+      given.container.addEventListener(`cocooned:after-${event}`, listener)
+      dispatch()
+
+      expect(listener).not.toHaveBeenCalled()
+    })
+
+    it.skip('is canceled if default behavior of the before event is prevented', () => {
+      const canceler = jest.fn(e => e.preventDefault())
+      $(given.container).on(`cocooned:before-${event}`, canceler)
+
+      const listener = jest.fn()
+      $(given.container).on(`cocooned:after-${event}`, listener)
+      trigger()
+
+      expect(listener).not.toHaveBeenCalled()
+    })
+    /* eslint-enable jest/no-identical-title */
+  })
+}