+++ npm/cocooned.js
@@ -1,21 +1,27 @@
 import $ from 'jquery'
-import { Cocooned as Base } from './src/cocooned'
+import { Base } from './src/cocooned/base'
 import { limitMixin, reorderableMixin } from './src/cocooned/plugins'
+import { CocoonSupportMixin } from './src/integrations/cocoon'
+import { jQueryPluginMixin } from './src/integrations/jquery'
 
-class Cocooned extends reorderableMixin(limitMixin(Base)) {}
-
-// Expose a jQuery plugin
-$.fn.cocooned = function (options) {
-  return this.each(function () {
-    if (typeof $(this).data('cocooned') !== 'undefined') {
+class Cocooned extends CocoonSupportMixin(reorderableMixin(limitMixin(Base))) {
+  static create (container, options) {
+    if ('cocooned' in container.dataset) {
       return
     }
 
-    return new Cocooned(this, options)
-  })
+    return new Cocooned(container, options)
+  }
+
+  static start () {
+    document.querySelectorAll('*[data-cocooned-options]').forEach(element => Cocooned.create(element))
+  }
 }
 
+// Expose a jQuery plugin
+jQueryPluginMixin($, Cocooned)
+
 // On-load initialization
-$(() => $('*[data-cocooned-options]').each((_i, el) => $(el).cocooned()))
+$(() => Cocooned.start())
 
 export default Cocooned