+++ npm/src/cocooned/base.js
@@ -1,9 +1,14 @@
 import $ from 'jquery'
-import Builder from './cocooned/builder'
+import Builder from './builder'
+import Emitter from './emitter'
 
-class Cocooned {
+class Base {
   static defaultOptions () {
     return {}
+  }
+
+  static eventNamespaces () {
+    return ['cocooned']
   }
 
   constructor (container, options) {
@@ -16,6 +21,24 @@ class Cocooned {
 
     this.init()
     this.container.get(0).dataset.cocooned = this
+  }
+
+  #emitter
+
+  get emitter () {
+    if (typeof this.#emitter === 'undefined') {
+      this.#emitter = new Emitter(this.constructor.eventNamespaces())
+    }
+
+    return this.#emitter
+  }
+
+  notify (node, eventType, eventData) {
+    if (node instanceof $) {
+      return this.emitter.emit(node.get(0), eventType, eventData)
+    }
+
+    return this.emitter.emit(node, eventType, eventData)
   }
 
   elementsCounter = 0
@@ -47,25 +70,12 @@ class Cocooned {
     return config
   }
 
-  notify (node, eventType, eventData) {
-    return !(this.namespaces.events.some(namespace => {
-      const namespacedEventType = [namespace, eventType].join(':')
-      const event = $.Event(namespacedEventType, eventData)
-      const eventArgs = [eventData.cocooned]
-
-      if ('node' in eventData) {
-        eventArgs.unshift(eventData.node)
-      } else if ('nodes' in eventData) {
-        eventArgs.unshift(eventData.nodes)
-      }
-
-      node.trigger(event, eventArgs)
-      return (event.isPropagationStopped() || event.isDefaultPrevented())
-    }))
+  selectors (type, selector = '&') {
+    return this.classes[type].map(klass => selector.replace(/&/, `.${klass}`))
   }
 
   selector (type, selector = '&') {
-    return this.classes[type].map(klass => selector.replace(/&/, `.${klass}`)).join(', ')
+    return this.selectors(type, selector).join(', ')
   }
 
   namespacedNativeEvents (type) {
@@ -152,26 +162,25 @@ class Cocooned {
     const self = this
 
     // Bind add links
-    this.addLinks.on(
-      this.namespacedNativeEvents('click'),
-      function (e) {
+    this.addLinks.each((_i, link) => {
+      link.addEventListener('click', function (e) {
         e.preventDefault()
-        e.stopPropagation()
-
         self.add(this, e)
       })
+    })
 
     // Bind remove links
     // (Binded on document instead of container to not bypass click handler defined in jquery_ujs)
-    $(document).on(
-      this.namespacedNativeEvents('click'),
-      this.selector('remove', `#${this.container.attr('id')} &`),
-      function (e) {
-        e.preventDefault()
-        e.stopPropagation()
+    document.addEventListener('click', function (e) {
+      const { target } = e
+      if (!self.selectors('remove').some(s => target.matches(s)) ||
+          target.closest(self.selector('container')) !== self.container.get(0)) {
+        return
+      }
 
-        self.remove(this, e)
-      })
+      e.preventDefault()
+      self.remove(target, e)
+    })
   }
 
   add (adder, originalEvent) {
@@ -240,5 +249,5 @@ class Cocooned {
 }
 
 export {
-  Cocooned
+  Base
 }